requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
certifi==2024.8.30
charset-normalizer==3.4.0
distro==1.9.0
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
idna==3.10
jiter==0.7.1
MouseInfo==0.1.3
openai==1.55.0
packaging==24.2
pillow==11.0.0
PyAutoGUI==0.9.54
pydantic==2.10.1
pydantic_core==2.27.1
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
pytesseract==0.3.13
python-dotenv==1.0.1
python3-xlib==0.15
pytweening==1.2.0
requests==2.32.3
sniffio==1.3.1
tqdm==4.67.0
typing_extensions==4.12.2
urllib3==2.2.3
